tests now also compile again
