+++ hgeometry/hgeometry.cabal
@@ -712,7 +712,6 @@ test-suite hspec
       ConvexHull.MelkmanSpec
       ClosestPair.ClosestPairSpec
       IntervalTreeSpec
-      LowerEnvelope.NaiveSpec
       LowerEnvelope.RegionsSpec
       Polygon.Convex.ConvexSpec
       Polygon.TriangulateSpec
