+++ hgeometry/src/HGeometry/VoronoiDiagram.hs
@@ -13,6 +13,9 @@ module HGeometry.VoronoiDiagram
   , VoronoiDiagram'(..)
   , voronoiDiagram
   , voronoiVertices
+  , asMap
+
+  , ColinearPoint(..)
   ) where
 
 import HGeometry.VoronoiDiagram.ViaLowerEnvelope
