+++ hgeometry/test-with-ipe/test/VoronoiDiagram/VoronoiSpec.hs
@@ -12,7 +12,6 @@ import           Golden
 import           HGeometry.Duality
 import           HGeometry.Ext
 import           HGeometry.Plane.LowerEnvelope
-import           HGeometry.Plane.LowerEnvelope.AdjListForm
 import qualified Data.Set as Set
 import qualified Data.List.NonEmpty as NonEmpty
 import qualified Data.Sequence as Seq
@@ -29,6 +28,7 @@ import           Test.Hspec.WithTempFile
 import           System.OsPath
 import           Test.Hspec
 import           Test.QuickCheck.Instances ()
+import           Plane.LowerEnvelopeSpec () -- imports the ipe instances for Voronoi Diagram
 -- import Test.Util
 
 --------------------------------------------------------------------------------
@@ -40,16 +40,18 @@ spec = describe "Voronoi diagram tests" $ do
     -- prop "voronoi vertex is center disk" $ \c ->
     --   voronoiVertices inputs
     it "vertices of a trivial voronoi diagram" $
-      voronoiVertices inputs `shouldBe` [Point2 5 5]
+      voronoiVertices inputs `shouldBe` (Set.fromList [Point2 5 5])
     -- it "a trivial voronoi diagram" $
     --   voronoiDiagram inputs `shouldBe` trivialVD
 
-    it "geometries of the trivial VD correct" $
-      trivialVD^..edgeGeometries
-      `shouldBe` [Left (HalfLine (Point2 5 5) (Vector2 1 0))
-                 ,Left (HalfLine (Point2 5 5) (Vector2 0 (-1)))
-                 ,Left (HalfLine (Point2 5 5) (Vector2 (-1) 1))
-                 ]
+    -- it "geometries of the trivial VD correct" $
+    --   trivialVD^..edgeGeometries
+    --   `shouldBe` [Left (HalfLine (Point2 5 5) (Vector2 1 0))
+    --              ,Left (HalfLine (Point2 5 5) (Vector2 0 (-1)))
+    --              ,Left (HalfLine (Point2 5 5) (Vector2 (-1) 1))
+    --              ]
+
+
     goldenWith [osp|data/test-with-ipe/golden/|]
                (ipeContentGolden { name = [osp|trivalVoronoi|] })
                  [ iO' inputs
@@ -102,52 +104,55 @@ grow d (Box p q) = Box (p&coordinates %~ subtract d)
                        (q&coordinates %~ (+d))
 
 
-instance (HasDefaultIpeOut point, Point_ point 2 r, Fractional r, Ord r
-         , Show r, Show point
-         )
-         => HasDefaultIpeOut (VoronoiDiagram point) where
-  type DefaultIpeOut (VoronoiDiagram point) = Group
-  defIO = \case
-    AllColinear _pts -> ipeGroup []
-    ConnectedVD vd  -> defIO vd
+-- instance (HasDefaultIpeOut point, Point_ point 2 r, Fractional r, Ord r
+--          , Show r, Show point
+--          )
+--          => HasDefaultIpeOut (VoronoiDiagram point) where
+--   type DefaultIpeOut (VoronoiDiagram point) = Group
+--   defIO = \case
+--     AllColinear _pts -> ipeGroup []
+--     ConnectedVD vd  -> defIO vd
 
 
-instance (HasDefaultIpeOut point, Point_ point 2 r, Fractional r, Ord r
-         , Show r, Show point
-         )
-         => HasDefaultIpeOut (VoronoiDiagram' point) where
-  type DefaultIpeOut (VoronoiDiagram' point) = Group
-  defIO vd = ipeGroup $ vd^..edgeGeometries.to render
-    where
-      bRect = boundingBox $ defaultBox :| [grow 1 $ boundingBox vd]
-      render = \case
-        Left hl   -> iO $ ipeHalfLineIn bRect hl
-        Right seg -> iO' seg
-
-inputs :: [Point 2 R]
-inputs = [origin, Point2 10 10, Point2 10 0]
+-- instance (HasDefaultIpeOut point, Point_ point 2 r, Fractional r, Ord r
+--          , Show r, Show point
+--          )
+--          => HasDefaultIpeOut (VoronoiDiagram' point) where
+--   type DefaultIpeOut (VoronoiDiagram' point) = Group
+--   defIO vd = ipeGroup $ vd^..edgeGeometries.to render
+--     where
+--       bRect = boundingBox $ defaultBox :| [grow 1 $ boundingBox vd]
+--       render = \case
+--         Left hl   -> iO $ ipeHalfLineIn bRect hl
+--         Right seg -> iO' seg
 
 
-trivialVD :: VoronoiDiagram' (Point 2 R)
-trivialVD = VoronoiDiagram $ LowerEnvelope vInfty (Seq.fromList [bv])
-  where
-    vInfty = UnboundedVertex $ Seq.fromList [Edge 1 h2 h3
-                                            ,Edge 1 h3 h1
-                                            ,Edge 1 h1 h2
-                                            ]
-    bv = Vertex (Point3 5 5 0)
-                (Set.fromList planes)
-                (Seq.fromList $
-                 [ Edge 0 h2 h1
-                 , Edge 0 h3 h2
-                 , Edge 0 h1 h3
-                 ]
-                )
-    planes = map (\p -> liftPointToPlane p :+ p) inputs
-    (h1,h2,h3) = case planes of
-                   [h1',h2',h3'] -> (h1',h2',h3')
-                   _             -> error "absurd"
-  -- order of the planes is incorrect, as is the z-coord.
+
+inputs :: NonEmpty (Point 2 R)
+inputs = NonEmpty.fromList [origin, Point2 10 10, Point2 10 0]
+
+trivialVD :: VoronoiDiagram (Point 2 R)
+trivialVD = voronoiDiagram inputs
+
+  -- VoronoiDiagram $ LowerEnvelope vInfty (Seq.fromList [bv])
+  -- where
+  --   vInfty = UnboundedVertex $ Seq.fromList [Edge 1 h2 h3
+  --                                           ,Edge 1 h3 h1
+  --                                           ,Edge 1 h1 h2
+  --                                           ]
+  --   bv = Vertex (Point3 5 5 0)
+  --               (Set.fromList planes)
+  --               (Seq.fromList $
+  --                [ Edge 0 h2 h1
+  --                , Edge 0 h3 h2
+  --                , Edge 0 h1 h3
+  --                ]
+  --               )
+  --   planes = map (\p -> liftPointToPlane p :+ p) inputs
+  --   (h1,h2,h3) = case planes of
+  --                  [h1',h2',h3'] -> (h1',h2',h3')
+  --                  _             -> error "absurd"
+  -- -- order of the planes is incorrect, as is the z-coord.
 
 
 testIpe            :: OsPath -> OsPath -> Spec
@@ -159,7 +164,7 @@ testIpe inFp outFp = do
         vv = voronoiVertices $ (view core) <$> points
         out = [ iO' points
               , iO' vd
-              ] <> [ iO'' v $ attr SStroke red | v <- vv ]
+              ] <> [ iO'' v $ attr SStroke red | v <- Set.toAscList vv ]
     goldenWith [osp|data/test-with-ipe/VoronoiDiagram/|]
                (ipeContentGolden { name = outFp })
                out
